Factor the new diagnostics log
